Move notification retrieval logic into perf platform client and dont filter by status
