+++ tests/app/clients/test_performance_platform.py
@@ -1,7 +1,15 @@
 import requests_mock
 import pytest
+from datetime import datetime
+from freezegun import freeze_time
+from functools import partial
 
 from app.clients.performance_platform.performance_platform_client import PerformancePlatformClient
+from app.utils import (
+    get_london_midnight_in_utc,
+    get_midnight_for_day_before
+)
+from tests.app.conftest import sample_notification_history as create_notification_history
 
 
 @pytest.fixture(scope='function')
@@ -69,3 +77,44 @@ def test_send_platform_stats_creates_correct_call(notify_api, client):
     assert request_args['count'] == 142
     expected_base64_id = 'MjAxNi0xMC0xNlQwMDowMDowMCswMDowMGdvdnVrLW5vdGlmeXNtc25vdGlmaWNhdGlvbnNkYXk='
     assert request_args['_id'] == expected_base64_id
+
+
+@freeze_time("2016-01-11 12:30:00")
+def test_get_total_sent_notifications_yesterday_returns_expected_totals_dict(
+    notify_db,
+    notify_db_session,
+    client,
+    sample_template
+):
+    notification_history = partial(
+        create_notification_history,
+        notify_db,
+        notify_db_session,
+        sample_template,
+        status='delivered'
+    )
+
+    notification_history(notification_type='email')
+    notification_history(notification_type='sms')
+
+    # Create some notifications for the day before
+    yesterday = datetime(2016, 1, 10, 15, 30, 0, 0)
+    with freeze_time(yesterday):
+        notification_history(notification_type='sms')
+        notification_history(notification_type='sms')
+        notification_history(notification_type='email')
+        notification_history(notification_type='email')
+        notification_history(notification_type='email')
+
+    total_count_dict = client.get_total_sent_notifications_yesterday()
+
+    assert total_count_dict == {
+        "start_date": get_midnight_for_day_before(datetime.utcnow()),
+        "end_date": get_london_midnight_in_utc(datetime.utcnow()),
+        "email": {
+            "count": 3
+        },
+        "sms": {
+            "count": 2
+        }
+    }
